feature: Codecc红线轮询修改 #988
